+++ bigbluebutton-html5/imports/startup/server/collection_methods/chat.js
@@ -1,5 +1,5 @@
 import { Chat } from '/collections/collections';
-import { logger } from '/server/server.js';
+import { logger } from '/imports/startup/server/logger';
 
 export function addChatToCollection(meetingId, messageObject) {
   let id;
