+++ bigbluebutton-html5/imports/startup/server/collection_methods/shapes.js
@@ -1,5 +1,5 @@
 import { Shapes, WhiteboardCleanStatus } from '/collections/collections';
-import { logger } from '/server/server.js';
+import { logger } from '/imports/startup/server/logger';
 
 export function addShapeToCollection(meetingId, whiteboardId, shapeObject) {
   let entry, id, removeTempTextShape;
