+++ bigbluebutton-html5/imports/startup/server/methods/muteUser.js
@@ -1,22 +1,22 @@
-import { publish } from '/server/redispubsub';
-import { isAllowedTo } from '/server/user_permissions';
-import { appendMessageHeader } from '/server/helpers';
-import { updateVoiceUser } from '/server/collection_methods/users';
-import { logger } from '/server/server';
+import { publish } from '/imports/startup/server/redispubsub';
+import { isAllowedTo } from '/imports/startup/server/user_permissions';
+import { appendMessageHeader } from '/imports/startup/server/helpers';
+import { updateVoiceUser } from '/imports/startup/server/collection_methods/users';
+import { logger } from '/imports/startup/server/logger';
 import { redisConfig } from '/config';
 
 Meteor.methods({
   // meetingId: the meetingId of the meeting the user[s] is in
-  // toMuteUserId: the userId of the user to be unmuted
+  // toMuteUserId: the userId of the user to be muted
   // requesterUserId: the userId of the requester
   // requesterToken: the authToken of the requester
-  unmuteUser(meetingId, toMuteUserId, requesterUserId, requesterToken) {
+  muteUser(meetingId, toMuteUserId, requesterUserId, requesterToken) {
     let action, message;
     action = function () {
       if (toMuteUserId === requesterUserId) {
-        return 'unmuteSelf';
+        return 'muteSelf';
       } else {
-        return 'unmuteOther';
+        return 'muteOther';
       }
     };
 
@@ -25,17 +25,17 @@ Meteor.methods({
         payload: {
           user_id: toMuteUserId,
           meeting_id: meetingId,
-          mute: false,
+          mute: true,
           requester_id: requesterUserId,
         }
       };
       message = appendMessageHeader('mute_user_request_message', message);
-      logger.info(`publishing a user unmute request for ${toMuteUserId}`);
+      logger.info(`publishing a user mute request for ${toMuteUserId}`);
       publish(redisConfig.channels.toBBBApps.users, message);
       updateVoiceUser(meetingId, {
         web_userid: toMuteUserId,
         talking: false,
-        muted: false,
+        muted: true,
       });
     }
   }
